+++ src/main/java/act/util/AsmTypes.java
@@ -69,6 +69,7 @@ public enum AsmTypes implements Opcodes {
     public static AsmType<TreeView> TREE_VIEW = new AsmType<>(TreeView.class);
     public static AsmType<act.cli.CsvView> CSV_VIEW_DEPRECATED = new AsmType<>(act.cli.CsvView.class);
     public static AsmType<CsvView> CSV_VIEW = new AsmType<>(CsvView.class);
+    public static AsmType<CommandPrefix> CMD_PREFIX = new AsmType<>(CommandPrefix.class);
 
     public static AsmType<SubClassFinder> SUB_CLASS_FINDER = new AsmType<>(SubClassFinder.class);
     public static AsmType<AnnotatedClassFinder> ANN_CLASS_FINDER = new AsmType<>(AnnotatedClassFinder.class);